Revert moving the ProductAutocompleteChoiceType
