+++ src/Sylius/Bundle/AdminBundle/Form/Type/Product/ProductAssociationsType.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Sylius\Bundle\AdminBundle\Form\Type\Product;
 
 use Doctrine\Common\Collections\Collection;
+use Sylius\Bundle\AdminBundle\Form\Type\ProductAutocompleteChoiceType;
 use Sylius\Bundle\ResourceBundle\Form\Type\FixedCollectionType;
 use Sylius\Component\Product\Model\ProductAssociationInterface;
 use Sylius\Component\Product\Model\ProductAssociationTypeInterface;
