+++ src/Sylius/Bundle/ProductBundle/Form/Type/ProductAutocompleteChoiceType.php
@@ -11,37 +11,38 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\AdminBundle\Form\Type\Product;
+namespace Sylius\Bundle\ProductBundle\Form\Type;
 
+use Sylius\Bundle\ResourceBundle\Form\Type\ResourceAutocompleteChoiceType;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\FormInterface;
+use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\UX\Autocomplete\Form\AsEntityAutocompleteField;
-use Symfony\UX\Autocomplete\Form\BaseEntityAutocompleteType;
 
-#[AsEntityAutocompleteField(route: 'sylius_admin_entity_autocomplete_admin')]
 final class ProductAutocompleteChoiceType extends AbstractType
 {
-    public function __construct (
-        private readonly string $productClass,
-    ) {
-    }
-
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            'class' => $this->productClass,
+            'resource' => 'sylius.product',
             'choice_name' => 'name',
             'choice_value' => 'code',
         ]);
     }
 
+    public function buildView(FormView $view, FormInterface $form, array $options): void
+    {
+        $view->vars['remote_criteria_type'] = 'contains';
+        $view->vars['remote_criteria_name'] = 'phrase';
+    }
+
     public function getBlockPrefix(): string
     {
-        return 'sylius_admin_product_autocomplete_choice';
+        return 'sylius_product_autocomplete_choice';
     }
 
     public function getParent(): string
     {
-        return BaseEntityAutocompleteType::class;
+        return ResourceAutocompleteChoiceType::class;
     }
 }
